requirements.txt
torch>=1.12.0
numpy>=1.21.0
matplotlib>=3.5.0
gym>=0.21.0
pyyaml>=6.0
Pillow>=9.0.0
